Cleaning up for Peakaboo 3

+++ Peakaboo_Dataset/src/peakaboo/dataset/provider/implementations/LocalDataSetProvider.java
@@ -137,7 +137,7 @@ public class LocalDataSetProvider extends DataSetProvider
 		}
 
 		float Net = (float) Ne / (float) Nt;
-		float Ntte = (float) Nt / ((float) Nt - (float) Ne);
+		float Ntte = Nt / ((float) Nt - (float) Ne);
 
 		Spectrum goodAverage = new Spectrum(dsc_average.size());
 		for (int i = 0; i < dsc_average.size(); i++)
